teams/meta-ads-sync/pyproject.toml
[project]
name = "meta-ads-sync"
version = "0.1.0"
description = "Multi-agent POC: Meta Ads daily spend → Google Sheets"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["lib"]
